fix: 좋아요 집계 이벤트 처리 Retryable 수정
