+++ apps/commerce-api/src/main/java/com/loopers/domain/product/event/ProductLikeEvent.java
@@ -1,4 +1,4 @@
-package com.loopers.domain.product;
+package com.loopers.domain.product.event;
 
 import java.util.UUID;
